+++ prebuild/tuple.ts
@@ -1,10 +1,9 @@
-import { commas, iota, joiner, txCommas } from './helpers';
+import { commas, iota, txCommas } from './helpers';
 
 const signatureFor = (num: number, opt?: boolean): string =>
   `
         /**
          * For tuples of [${txCommas(num)}]
-         * ${joiner(num, v => `@param arb${v} Arbitrary responsible for T${v}`, '\n* ')}
          */
         function tuple<${txCommas(num)}>(
                 ${commas(num, v => `arb${v}${opt === true ? '?' : ''}: Arbitrary<T${v}>`)}